+++ mipidsi/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - replaced `window_offset_handler` function pointer with `offset` field
 - default to disabled color inversion for all generic models
 - renamed `Display::set_scroll_region` and `Display::set_scroll_offset` into `set_vertical_scroll_region` and `set_vertical_scroll_offset`
+- removed `Model::default_options`
 
 ### Removed